% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ampute3.R
\name{ampute3}
\alias{ampute3}
\title{generate missing data on complete or incomplete data under MAR, MCAR and MNAR mechanisms and patterns}
\usage{
ampute3(
  data,
  mechanism = "MCAR",
  perc.missing = 0.2,
  self.mask = NULL,
  idx.incomplete = NULL,
  idx.covariates = NULL,
  weights.covariates = NULL,
  by.patterns = FALSE,
  patterns = NULL,
  freq.patterns = NULL,
  weights.patterns = NULL,
  use.all = FALSE,
  logit.model = "RIGHT",
  seed = 1234
)
}
\arguments{
\item{data}{\link{data.frame, matrix} a complet of incomplete data frame of matrix.}

\item{mechanism}{\link{string} could be "MCAR", "MAR" or "MNAR". Default is "MCAR".}

\item{perc.missing}{\link{positive double} proportion of missing values, between 0 and 1. Default is 0.5.}

\item{self.mask}{\link{string} either NULL or one of "sym", "upper", "lower"; default is NULL}

\item{idx.incomplete}{\link{array} indices of variables to generate missing values in. If NULL then missing values in all variables are possible. Default is NULL.}

\item{idx.covariates}{\link{matrix} binary matrix such that entries in row i that are equal to 1 indicate covariates that incluence missingness of variable i (sum(idx.incomplete) x p). If NULL all covariates contribute. Default is NULL.}

\item{weights.covariates}{\link{matrix} matrix of same size as idx.covariates with weights in row i for contribution of each covariate to missingness model of variable i. If NULL then a (regularized) logistic model is fitted. Default is NULL.}

\item{by.patterns}{\link{boolean} generate missing values according to (pre-specified) patterns. Default is FALSE.}

\item{patterns}{\link{matrix} binary matrix with 1=observed, 0=missing (n_pattern x p). Default is NULL.}

\item{freq.patterns}{\link{array} array of size n_pattern containing desired proportion of each pattern. if NULL then mice::ampute.default.freq will be called. Default is NULL.}

\item{weights.patterns}{\link{matrix} weights used to calculate weighted sum scores (n_pattern x p). If NULL then mice::ampute.default.weights will be called. Default is NULL.}

\item{use.all}{\link{boolean} use all observations, including incomplete observations, for amputation when amputing by patterns (only relevant if initial data is incomplete and by.pattern=T). Default is FALSE.}

\item{logit.model}{\link{string} either one of "RIGHT","LEFT","MID","TAIL". Default is "RIGHT".}

\item{seed}{\link{natural integer} seed for random numbers generator. Default is 1234.}
}
\value{
A list with the following elements
\item{data.init}{original data.frame}
\item{data.incomp}{data.frame with the newly generated missing values, observed values correspond to the values from the initial data.frame}
\item{idx_newNA}{a boolean data.frame indicating the indices of the newly generated missing values}
}
\description{
generate missing data on complete or incomplete data under MAR, MCAR and MNAR mechanisms and patterns
}
\references{
Schouten, R.M., Lugtig, P and Vink, G. (2018) \href{https://www.tandfonline.com/doi/full/10.1080/00949655.2018.1491577}{Generating missing values for simulation purposes: A multivariate amputation procedure.}. \emph{Journal of Statistical Computation and Simulation}, 88(15): 1909-1930.
Teresa Alves de Sousa et Imke Mayer (2021) \href{https://rmisstastic.netlify.app/how-to/generate/misssimul#311_MCAR}{How to generate missing values?}
}
