% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_coxpred.R
\name{coxpred}
\alias{coxpred}
\title{Title}
\usage{
coxpred(data, R = 5, seed = 1234, model = "fullcox", ...)
}
\arguments{
\item{data}{}

\item{R}{}

\item{seed}{}

\item{model}{}

\item{...}{}
}
\value{

}
\description{
Title
}
