% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_coxest.R
\name{evaluate_coxest}
\alias{evaluate_coxest}
\title{Title}
\usage{
evaluate_coxest(estimate, true_val, metric = "bias")
}
\arguments{
\item{estimate}{}

\item{true_val}{}

\item{metric}{}
}
\value{

}
\description{
Title
}
