% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_coxmodel.R
\name{full_coxmodel}
\alias{full_coxmodel}
\title{Title}
\usage{
full_coxmodel(fullcox, myformula, scenario = "baseline", method = "fullcox")
}
\arguments{
\item{fullcox}{}

\item{myformula}{}

\item{scenario}{}

\item{method}{}
}
\value{

}
\description{
Title
}
