% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencox.R
\name{gencox}
\alias{gencox}
\title{Generate survival data from a Cox model}
\usage{
gencox(
  N = 1000,
  beta = rep(0.5, 5),
  lambda = 0.002,
  rateC = 0.004,
  rho = 0.003
)
}
\arguments{
\item{N}{Sample size}

\item{beta}{true beta vector}

\item{lambda}{baseline hazard (scale parameter in h0)}

\item{rateC}{rate parameter of the exponential distribution of Censoring}

\item{rho}{scale parameter}
}
\value{
data.frame with time (survival time), status(1 = censored), predictors (x_i)
time(survival time), cumhaz (cumulative hazad),
}
\description{
Generate survival data with $beta$, $lambda_t$ baseline of censoring and $labmda_c$ hazard of censoring
}
\references{
N.J. Horton and Ken K.Keinman, Using R and RStudio for Data Management and Statistical Analysis and Graphics, second edition, 2015
}
