% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cox.R
\name{generate_cox}
\alias{generate_cox}
\title{Generate survival data from a Cox model}
\usage{
generate_cox(n = 1000, beta = rep(0.5, 5), lambdaT = 0.002, lambdaC = 0.004)
}
\arguments{
\item{n}{Sample size}

\item{beta}{true beta}

\item{lambdaT}{baseline hazard}

\item{lambdaC}{hazard of censoring}

\item{x}{covariate}
}
\value{
data.frame with time (survival time), event(0 = censored), predictors (x_i)
T (survival time), event (0: censored),
}
\description{
Generate survival data with $beta$, $lambda_t$ baseline of censoring and $labmda_c$ hazard of censoring
}
\references{
N.J. Horton and Ken K.Keinman, Using R and RStudio for Data Management and Statistical Analysis and Graphics, second edition, 2015
}
