% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateNA.R
\name{generateNA}
\alias{generateNA}
\title{Title}
\usage{
generateNA(x, p = 0.2, seed = 123)
}
\arguments{
\item{x}{dataset with NA}

\item{p}{proportion of NA}

\item{seed}{seed value}
}
\value{
value
}
\description{
Title
}
