\name{impute-package}
\alias{impute-package}
\docType{package}
\title{
Imputes missing data
}
\description{
Imputes missing data using mean, median, mode or using classifer or regression.
}
\details{
\tabular{ll}{
Package: \tab impute\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-22\cr
}
Functions: impute, predict, inject.na
}
\author{
Piotr Takiel
}
\examples{

}
