% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputed_coxmodel.R
\name{imputed_coxmodel}
\alias{imputed_coxmodel}
\title{Title}
\usage{
imputed_coxmodel(
  impcox,
  myformula,
  method = NULL,
  scenario = NULL,
  truelogHR = logBeta
)
}
\arguments{
\item{impcox}{}

\item{myformula}{}

\item{method}{}

\item{scenario}{}

\item{truelogHR}{}
}
\description{
Title
}
