% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R, R/package.R
\docType{package}
\name{imputer}
\alias{imputer}
\title{Impute mixed-type data with missing values using a specific algorithm}
\usage{
imputer(data_with_NA, "imputation_method")
}
\arguments{
\item{data}{A data frame or matrix containing missing values}

\item{method}{Imputation algorithms wrapped in the imputer function : "naive", "hotdeck", "knn", "cart", "missranger", "missforest", "spmm", "famd", "mpmm", "micerf", "micecart",}
}
\value{
A single complete data set imputed with the chosen method.
\code{imputer} is a wrapper function that gives a direct use of the specified imputation method. All the available methods support mixed data type (continuous and categroical).
\code{"naive"} imputation by mean for continuous and mode for categorical variables
\code{"hotdeck"}  single sequential hot deck imputation
\code{"knn"}  single imputation by k-nearest neighbor
\code{"cart"} single imputation by classification and regression trees
\code{"missforest"}  single nonparametric imputation using random forest algorithm
\code{"spmm"}  single predictive mean matching
\code{"famd"}  single imputation by factorial analysis for mixed-type
\code{"missranger"} single imputation  by chained random forests
\code{"misscforest"} single imputation  by ensemble conditional trees
\code{"mice"}  multiple imputation by chained equations
\code{"mpmm"} multiple predictive mean matching
\code{"micerf"}  multiple Imputation by random forests
\code{"micecart"}  multiple  imputation by classification and regression #
\code{"mixgb"}  multiple imputation by xgboost
}
\description{
add description

A consistent wrapper for a set of imputation algorithms
}
\examples{
# add a simple example here to introduce the package!

}
