% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introduceNA.R
\name{introduceNA}
\alias{introduceNA}
\title{Title}
\usage{
introduceNA(x, p = 0.2, seed = 123)
}
\arguments{
\item{x}{dataset with NA}

\item{p}{proportion of NA}

\item{seed}{seed value}
}
\value{
value
}
\description{
Title
}
\examples{
irisNA <- introduceNA(iris, 0.2)
mean(is.na(iris))
}
