% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_na.R
\name{summarise_na}
\alias{summarise_na}
\title{Title}
\usage{
summarise_na(data)
}
\arguments{
\item{data}{}
}
\description{
Title
}
\examples{
#----------- summarise_na() : summarise NA by cases and variables (wrapper for naniar::miss_var_summary & naniar::miss_case_summary)
}
