% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableNA.R
\name{tableNA}
\alias{tableNA}
\title{Title}
\usage{
tableNA(data, order = TRUE, cumsum = FALSE)
}
\arguments{
\item{data}{a dataframe object}

\item{order}{sort the columns based on the proportion of NA}

\item{cumsum}{compute and display the cumulative sum of NA across all columns}
}
\value{
value
}
\description{
Title
}
\examples{
tableNA(introduceNA(iris, 0.2))
}
