% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_coxest.R
\name{evaluate_coxest}
\alias{evaluate_coxest}
\title{Title}
\usage{
evaluate_coxest(estimate, truelogHR, method = "naive")
}
\arguments{
\item{estimate}{}

\item{truelogHR}{}

\item{evaluate_coxest(c(5, }{5, 0.6), c(7, 0.6, 0.3))}
}
\value{
data frame containing results
}
\description{
Title
}
\examples{
evaluate_coxest(c(5, 5, 0.6), c(7, 0.6, 0.3)) 
}
