% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_coxpred.R
\name{evaluate_coxpred}
\alias{evaluate_coxpred}
\title{Title}
\usage{
evaluate_coxpred(dat, myformula, metric, seed = 1234, ...)
}
\arguments{
\item{dat}{}

\item{myformula}{}

\item{metric}{AUC, IBS or predError}

\item{seed}{radnom number generation}

\item{...}{}
}
\description{
Title
}
