% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_coxpred.R
\name{evaluate_coxpred}
\alias{evaluate_coxpred}
\title{Title}
\usage{
evaluate_coxpred(
  data,
  R = 5,
  seed = 1234,
  method = "fullcox",
  scenario = "MAR-10",
  ...
)
}
\arguments{
\item{data}{generate with the substantive cox model either complete or imputed}

\item{R}{number of replcations}

\item{seed}{radnom number generation}

\item{method}{}

\item{scenario}{}

\item{...}{}
}
\value{

}
\description{
Title
}
