% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputer.R, R/package.R
\docType{package}
\name{imputer}
\alias{imputer}
\title{impute variables with missing data values using a specific algorithm}
\usage{
imputer(dataNA, "method")
}
\arguments{
\item{data}{A data frame or matrix containing missing values}

\item{method}{Imputation algorithms wrapped in the imputer function : "naive", "hotdeck", "knn", "cart", "glmnet", "missranger", "missforest", "spmm", "famd", "mpmm", "micerf", "supermice".}
}
\value{
A single complete data set imputed with the chosen method.
}
\description{
A consistent wrapper for a set of imputation algorithms
}
\details{

}
\examples{

# add a simple example here to introduce the package!

}
\references{
evaluate_coxest(c(5, 5, 0.6), c(7, 0.6, 0.3))
}
