% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableNA.R
\name{tableNA}
\alias{tableNA}
\title{Title}
\usage{
tableNA(data, order = TRUE, cumsum = FALSE)
}
\arguments{
\item{data}{}
}
\description{
Title
}
\examples{
#----------- summarise_na() : summarise NA by cases and variables (wrapper for naniar::NA_var_summary & naniar::NA_case_summary)
}
